libs_and_topics/finance/part_time_larry/binance_tutorials/requirements.txt
# pip install -r requirements.txt
python-binance
TA-Lib
numpy
backtrader
flask
matplotlib
websocket_client
